+++ odoo/local-src/sf_rma_mrp_repair/__manifest__.py
@@ -2,8 +2,8 @@
 # Copyright 2017 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 {
-    "name": "Sensefly MRP",
-    "summary": "MRP related customizations",
+    "name": "Sensefly RMA MRP Repair",
+    "summary": "RMA MRP Repair related customizations",
     "version": "10.0.1.0.0",
     "category": "Manufacturing",
     "website": "https://www.camptocamp.com/",