+++ odoo/local-src/sf_account/models/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+# Part of sensefly.
+import account_journal