+++ odoo/local-src/sf_partner/models/partner.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+# Part of SenseFly.
+from odoo import fields, models
+
+
+class Partner(models.Model):
+    _inherit = 'res.partner'
+
+    is_intragroup = fields.Boolean(string='Is Intragroup',
+                                   help='This partner belongs to Parrot Group')