+++ HISTORY.rst
@@ -20,7 +20,7 @@ latest (unreleased)
 +++++++++++++++++++
 
 **Features and Improvements**
-
+* Add field owner in account analytic
 * Add Partner: Customer, supplier (draft)
 * Add the import Sales Team / Channel to demo and install songs
 * Add the import "Drone type" to demo and install songs