+++ HISTORY.rst
@@ -21,6 +21,7 @@ latest (unreleased)
 
 **Features and Improvements**
 
+* Add active field on journal
 * Add Partner: Customer, supplier (draft)
 * Add the import Sales Team / Channel to demo and install songs
 * Add the import "Drone type" to demo and install songs